refactor: factoring out the lexer
